requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0
openai>=1.0.0
python-dateutil>=2.8.0
pytz>=2023.3
requests>=2.31.0
flask>=2.3.0

